refactor: enhance type handling and error checking in API clients
- Updated function signatures in EdinetAPIClient and JQuantsAPIClient to accept optional API keys.
- Improved error handling in API request methods to raise specific exceptions.
- Adjusted return types in CompanyInfoService and DataService to use StockSearchResult for better type clarity.
- Refined type definitions in portfolio-related services to ensure correct data types for financial calculations.

+++ mebuki/utils/portfolio_types.py
@@ -68,5 +68,5 @@ class SectorAllocation(TypedDict):
     sector_name: str
     ticker_count: int
     tickers: list[str]
-    total_cost: int
+    total_cost: float
     ratio: float
